window functions: Fix tiebreaker in `order_aggregate_datums_with_rank`
Like everywhere else, the tiebreaker here should yield
`Ordering::Equal` only when the entire rows are equal. This
is to prevent non-determinism if the input order is
non-deterministic. (Btw. the input order here is probably
usually deterministic, but this function is called from many
places, so it's hard to be sure.)

Note that this changes the results in some slts, but this is fine,
because SQL doesn't guarantee an order for ordered aggregations
(unless the user explicitly specifies an ORDER BY in the aggregation
itself). For example, `array_agg` as a window aggregation used to
inherit the ordering from the OVER clause when it ddidn't have its
own ORDER BY, but this was just an accidental ordering.