Improve method names around lateral joins
The fn name `is_correlated` made me think that it checks whether
there is a col ref to _any_ outer scope (because that is what a
"correlated subquery" means). But actually, this fn is just for
sniffing out lateral joins: whether there is a col ref to the
outer scope directly above us (e.g., from the rhs of a join to
the lhs).

Also renamed `can_be_correlated` for consistency.