Refactor `create_fast_path_plan`
This commit separates an inner part, which
- doesn't need a full-blown DataflowDescription, only an
  MirRelationExpr;
- can either consider indexes from all clusters, or only the given
  indexes;
- returns a list of potentially several FastPathPlans.

The next commit will rely on the inner part in
PlanInsights::compute_fast_path_clusters to avoid iterating through
all clusters.

Also, this refactoring will be useful to avoid running
optimize_mir_local for fast path peeks. (This will be a follow-up
PR.)